+++ kostyukov.mikhail/common/scopeGuard.cpp
