+++ kostyukov.mikhail/common/scopeGuard.hpp
